Updated Contacts Page and Dashboard page
Fixed the contacts page. It should be working correctly. Added some routing portions to the dashboard page for easy access for the user. There is a small error where when multiple tabs are open and refreshed it becomes the same user logged in. It has something to do with not having a unique key for local storage.